As spotify URI format has changed, and sometimes it adds a "?si=string" after the ID (not sure what it does, but if you try to use it in the original program, it throws invalid spotify link error).
This is a lazy fix, for sure.

Removed "|| IsSpotifyPlaylistUrl(str)" from IsSpotifyURL
Removed IsSpotifyPlaylistURL class as it is no longer needed
Changed IsSpotifyTrackAlbumURL to include playlist (though it won't fix the issue of being able to download playlists until the package used is fixed to support playlists again)
Added a second delimiter to url.split(), "?" - This way we can remove the sometimes added extra string after the ID of playlist/album/track.