Patch eviabs#33
https://github.com/eviabs/downtify-premium/pull/33

Lovemonkeyz:

As spotify URI format has changed, and sometimes it adds a "?si=string" after the ID (not sure what it does, but if you try to use it in the original program, it throws invalid spotify link error). It won't fix album downloads until the package used is updated, but at least it removes any issues where a copied link contains the ?si= part.

This is a lazy fix, for sure.

    Removed "|| IsSpotifyPlaylistUrl(str)" from IsSpotifyURL
    Removed IsSpotifyPlaylistURL class as it is no longer needed
    Changed IsSpotifyTrackAlbumURL to include playlist
    Added a second delimiter to url.split(), "?" - This way we can remove the sometimes added extra string after the ID of playlist/album/track.